requirements.txt
django
djangorestframework
django-cors-headers
markdown      
django-filter 